client: finish traffic pattern generator
It is now possible to specify the desired bandwidth in any SI/IEC unit. Via

./client -4 -t udp -r 0 -R 530:1470:10kB -e 127.0.0.1

is is possible to generate IP packets (here IPv4) with a payload length
between 530 and 1470 byte. In average this is an 1000 byte packet. The
bandwidth - the third argument to R - specify the desired bandwidth. Here
10.000 byte per seconds. Which will result in an average packet
transmission delay of 100000 us (0.1 second).

Last but not least this commit extend the README file to explain the
above enhancement.

Signed-off-by: Hagen Paul Pfeifer <hagen@jauu.net>